+++ src/Servers/Kestrel/Core/src/Internal/Infrastructure/KestrelConnectionOfT.cs
@@ -8,35 +8,6 @@ using Microsoft.Extensions.Logging;
 
 namespace Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Infrastructure;
 
-internal interface IConnectionMetricsContextFeature
-{
-    ConnectionMetricsContext MetricsContext { get; }
-}
-
-internal readonly struct ConnectionMetricsContext
-{
-    public BaseConnectionContext ConnectionContext { get; }
-    public bool CurrentConnectionsCounterEnabled { get; }
-    public bool ConnectionDurationEnabled { get; }
-    public bool QueuedConnectionsCounterEnabled { get; }
-    public bool QueuedRequestsCounterEnabled { get; }
-    public bool CurrentUpgradedRequestsCounterEnabled { get; }
-    public bool CurrentTlsHandshakesCounterEnabled { get; }
-
-    public ConnectionMetricsContext(BaseConnectionContext connectionContext, bool currentConnectionsCounterEnabled,
-        bool connectionDurationEnabled, bool queuedConnectionsCounterEnabled, bool queuedRequestsCounterEnabled,
-        bool currentUpgradedRequestsCounterEnabled, bool currentTlsHandshakesCounterEnabled)
-    {
-        ConnectionContext = connectionContext;
-        CurrentConnectionsCounterEnabled = currentConnectionsCounterEnabled;
-        ConnectionDurationEnabled = connectionDurationEnabled;
-        QueuedConnectionsCounterEnabled = queuedConnectionsCounterEnabled;
-        QueuedRequestsCounterEnabled = queuedRequestsCounterEnabled;
-        CurrentUpgradedRequestsCounterEnabled = currentUpgradedRequestsCounterEnabled;
-        CurrentTlsHandshakesCounterEnabled = currentTlsHandshakesCounterEnabled;
-    }
-}
-
 internal sealed class KestrelConnection<T> : KestrelConnection, IThreadPoolWorkItem where T : BaseConnectionContext
 {
     private readonly Func<T, Task> _connectionDelegate;
